Move the internal impl in api/internal.rs
